requirements.txt
fastapi
uvicorn
python-multipart
pandas
python-pptx
openpyxl